# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * API specification for the Games Tracker
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * played = finished/done. playing = currently playing. on_hold = paused for later. dropped = abandoned.

 */
export type GameStatus = typeof GameStatus[keyof typeof GameStatus];


export const GameStatus = {
  played: 'played',
  playing: 'playing',
  on_hold: 'on_hold',
  dropped: 'dropped',
} as const;

export interface Game {
  id: string;
  title: string;
  platform?: string | null;
  genre?: string | null;
  status: GameStatus;
  /**
     * @minimum 1
     * @maximum 10
     */
  rating?: number | null;
  coverUrl?: string | null;
  notes?: string | null;
  hoursPlayed?: number | null;
  startedAt?: string | null;
  finishedAt?: string | null;
  steamAppId?: number | null;
  earnedAchievements?: string[];
  /** Absolute path to the game executable for launch and playtime tracking */
  executablePath?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateGameInput {
  /** @minLength 1 */
  title: string;
  platform?: string | null;
  genre?: string | null;
  status: GameStatus;
  /**
     * @minimum 1
     * @maximum 10
     */
  rating?: number | null;
  coverUrl?: string | null;
  notes?: string | null;
  hoursPlayed?: number | null;
  startedAt?: string | null;
  finishedAt?: string | null;
  executablePath?: string | null;
}

export interface UpdateGameInput {
  /** @minLength 1 */
  title?: string;
  platform?: string | null;
  genre?: string | null;
  status?: GameStatus;
  /**
     * @minimum 1
     * @maximum 10
     */
  rating?: number | null;
  coverUrl?: string | null;
  notes?: string | null;
  hoursPlayed?: number | null;
  startedAt?: string | null;
  finishedAt?: string | null;
  executablePath?: string | null;
}

export interface CoverSearchResult {
  coverUrl: string | null;
  title?: string | null;
  steamAppId?: number | null;
}

export interface Achievement {
  name: string;
  displayName: string;
  description?: string | null;
  iconUrl?: string | null;
  earned: boolean;
}

export type AchievementListSource = typeof AchievementListSource[keyof typeof AchievementListSource];


export const AchievementListSource = {
  steam: 'steam',
  none: 'none',
} as const;

export interface AchievementList {
  steamAppId?: number | null;
  total: number;
  earnedCount: number;
  achievements: Achievement[];
  source: AchievementListSource;
}

export interface ToggleAchievementInput {
  name: string;
  earned: boolean;
}

export type GameStatsTopGenresItem = {
  genre: string;
  count: number;
};

export type GameStatsTopPlatformsItem = {
  platform: string;
  count: number;
};

export interface GameStats {
  total: number;
  played: number;
  playing: number;
  onHold: number;
  dropped: number;
  totalHours: number;
  averageRating?: number | null;
  topGenres: GameStatsTopGenresItem[];
  topPlatforms: GameStatsTopPlatformsItem[];
}

export interface FriendsProfile {
  id: string;
  referralCode: string;
  displayName: string;
  friendCount: number;
}

export interface UpdateFriendsProfileInput {
  /** @minLength 1 */
  displayName: string;
}

export interface FriendSummary {
  id: string;
  referralCode: string;
  displayName: string;
}

export interface FriendsList {
  friends: FriendSummary[];
}

export interface InviteFriendInput {
  /** @minLength 4 */
  code: string;
}

export interface InviteFriendResult {
  friend: FriendSummary;
}

export interface FriendGameSummary {
  id: string;
  title: string;
  coverUrl?: string | null;
  platform?: string | null;
  status: string;
  rating?: number | null;
  hoursPlayed?: number | null;
  earnedAchievements: string[];
  steamAppId?: number | null;
}

export interface FriendRecommendation {
  friendId: string;
  friendName: string;
  game: FriendGameSummary;
  reason: string;
}

export interface SharedFriendStats {
  friendId: string;
  friendName: string;
  hoursPlayed?: number | null;
  earnedCount: number;
  status: string;
  rating?: number | null;
}

export interface SharedGameComparison {
  gameTitle: string;
  coverUrl?: string | null;
  steamAppId?: number | null;
  myGame: FriendGameSummary;
  friends: SharedFriendStats[];
}

export interface FriendRecommendations {
  friends: FriendSummary[];
  friendRecommendations: FriendRecommendation[];
  sharedComparisons: SharedGameComparison[];
}

export type SearchGameCoverParams = {
title: string;
};

export type ListGamesParams = {
status?: GameStatus;
search?: string;
};