# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for the Games Tracker
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /friends/profile:
    get:
      operationId: getFriendsProfile
      tags: [friends]
      summary: Get your referral code and friend count
      responses:
        "200":
          description: Profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FriendsProfile"
        "403":
          description: Sign in required
...
  /friends/invite:
    post:
      operationId: inviteFriend
      tags: [friends]
      summary: Add a friend using their referral code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InviteFriendInput"
      responses:
        "201":
          description: Friend added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteFriendResult"
        "404":
          description: Invalid code
...
  /friends/{friendId}:
    parameters:
      - in: path
        name: friendId
        required: true
        schema:
          type: string
    delete:
      operationId: removeFriend
      tags: [friends]
      summary: Remove a friend
      responses:
        "204":
          description: Removed
components:
  schemas:
...
    FriendsProfile:
      type: object
      properties:
        id:
          type: string
        referralCode:
          type: string
        displayName:
          type: string
        friendCount:
          type: integer
      required: [id, referralCode, displayName, friendCount]
    FriendSummary:
      type: object
      properties:
        id:
          type: string
        referralCode:
          type: string
        displayName:
          type: string
      required: [id, referralCode, displayName]
...
    InviteFriendInput:
      type: object
      properties:
        code:
          type: string
          minLength: 4
      required: [code]
    InviteFriendResult:
      type: object
      properties:
        friend:
          $ref: "#/components/schemas/FriendSummary"
      required: [friend]
...