# lib/api-zod/src/generated/types/inviteFriendInput.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Api
 * API specification for the Games Tracker
 * OpenAPI spec version: 0.1.0
 */

export interface InviteFriendInput {
  /** @minLength 4 */
  code: string;
}